Set up API, Redux and Axios
created an api.js file to hold the api url. added axios to the gamesAction file to get popular games. The API key is in a .env file and should not be visible to the public on github. Created an index.js file in the reducers folder to hold all the different reducers in the store. Added useEffect() hook to run loadGames action and fetch data when the component loads