stopped background from scrolling on game detail. And allowed toggles back to homepage when shadowcard is clicked
added a hidden overflow when the Game is clicked to open game details. This gets rid of the scroll on the Home page. Added a class to the background overlay of the game detail. Whenever that element is clicked the useNavigate function runs and brings the user back to the homepage, removing the game detail component