added icons
added platform icons. When the game details component loads there is a switch and case statement that will populate an icon svg based on the value. Also added star rating. Rounds the gameData.rating so it only shows full stars. Runs a for loop of 5 since the rating is out of 5. Creates a full star based on rating value and creates empty stars for the remaing.