added navbar and search input
because its a rest API there is another link to populate the games search based on name. Created another case in the gamesReducer to add the data to the searched key. Included the searched key in the home page useSelector to generate data from gamesreducer when there is values in the array and to create an empty string if there is no value. The input is added to state and is passed as the value when submited and dispatched. Clicking on LOGO clears the data stored in the games reducer search key and removes the searched games